\docType{data}
\name{CLAYr}
\alias{CLAYr}
\title{The Digital Soil Map of Sweden -topsoil clay content}
\format{Raster layer}
\usage{
data(CLAYr)
}
\description{
Excerpt from the Digital Soil Map of Sweden. Projected coordinate system: Sweref99TM (epsg: 3006). Attribute: predicted topsoil (0-20 cm depth)
clay content (% of fine soil).
}
\references{
Available online: https://www.sgu.se/samhallsplanering/planering-och-markanvandning/markanvandning/jordbruk-skog-och-fiske/lerhaltskartan-digital-akermarkskarta/.
}
\keyword{datasets}
