% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluate.R
\name{evaluate}
\alias{evaluate}
\title{evaluate}
\usage{
evaluate(df, observed, predicted)
}
\arguments{
\item{df}{Data.frame. Required. A data.frame with observed and predicted data.}

\item{observed}{Character value. Required. The  name of the column in df
with predicted data.The data must be of class numeric.}

\item{predicted}{Character value or vector. Required. The  names of the
column(s) in df with predicted data. The data  must be of class numeric.}
}
\value{
A data.frame with evaluation statistics. For details, see mri function.
}
\description{
Computes evaluation measures from observed and predicted data.
}
\examples{
df<-data.frame(obs=1:9, pred=c(2, 9, 10, 8, 3, 4, 6, 12, 1))
e<-evaluate(df, 'obs', 'pred')
print(e)

}
