% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/even.R
\name{even}
\alias{even}
\title{even}
\usage{
even(x)
}
\arguments{
\item{x}{Integer.}
}
\value{
Logical value (TRUE or FALSE). TRUE means that the value is even.
}
\description{
Checks whether an integer is even.
}
\examples{
even(3)

}
