% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extentpolygon.r
\name{extentpolygon}
\alias{extentpolygon}
\title{extentpolygon}
\usage{
extentpolygon(x)
}
\arguments{
\item{x}{A spatial object.}
}
\value{
SpatVector of polygons.
}
\description{
Create a SpatVector of polygons from the extent of a spatial object.
}
\details{
If x is projected, the SpatVector will also be projected
}
