% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feedback.R
\name{fback}
\alias{fback}
\title{fback}
\usage{
fback(t)
}
\arguments{
\item{t}{Character string. Required.}
}
\value{
A Character vector.
}
\description{
Prints feedback and adds it to a vector.
}
\keyword{internal}
