% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MAE.R
\name{mae}
\alias{mae}
\title{Mean Absolute Error (MAE)}
\usage{
mae(observed, predicted)
}
\arguments{
\item{observed}{a numeric vector of observed values}

\item{predicted}{a numeric vector of predicted values. The length shall be the same as for observed.}
}
\value{
the mean absolute error (MAE) calculated from the observed and the predicted values.
}
\description{
Calculates the mean absolute error (MAE) from observed and predicted values.
}
\details{
mae = mean(abs(observed - predicted))
}
\examples{
o<-1:5
p<-c(2,2,4,3,5)
mae(observed=o, predicted=p)
}
