% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ME.R
\name{me}
\alias{me}
\title{Mean Error (ME)}
\usage{
me(observed, predicted)
}
\arguments{
\item{observed}{a numeric vector of observed values}

\item{predicted}{a numeric vector of predicted values. The length shall be the same as for observed.}
}
\value{
the mean  error (ME) calculated from the observed and the predicted values.
}
\description{
Calculates the mean error (ME) from observed and predicted values.
}
\details{
me = bias = mean(observed - predicted)
}
\examples{
o<-1:5
p<-c(2,2,4,3,5)
me(observed=o, predicted=p)
}
