% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/odd.R
\name{odd}
\alias{odd}
\title{even}
\usage{
odd(x)
}
\arguments{
\item{x}{Integer.}
}
\value{
Logical value (TRUE or FALSE). TRUE means that the value is odd.
}
\description{
Checks whether an integer is odd.
}
\examples{
odd(3)

}
