% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spdf_from_extent.r
\name{spdf_from_extent}
\alias{spdf_from_extent}
\title{spdf_from_extent}
\usage{
spdf_from_extent(x)
}
\arguments{
\item{x}{A spatial object.}
}
\value{
SpatialPolygonsDataFrame.
}
\description{
Create a spatialPolygonsDataFrame from extent of a spatial object.
}
\details{
If x is projected, the spatialpolygonsdataframe will also be projected
}
\examples{
require(raster) #load required package.
r1<-raster::raster(ext=extent(c(0,10,0,10)), res=1, vals=1:100) #create example raster.
spdf<-spdf_from_extent(r1) #convert the raster extent to SpatialPolygonsdataFrame.
plot(spdf) #Plot results.

}
