% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/me.R
\name{me}
\alias{me}
\title{me}
\usage{
me(observed, predicted)
}
\arguments{
\item{observed}{Numeric vector of observed values}

\item{predicted}{Numeric vector of predicted values. The length shall be
the same as for observed.}
}
\value{
The mean error (ME) calculated from the observed and the predicted
values.
}
\description{
Calculates the mean error (ME) from observed and predicted values.
}
\details{
ME = bias = mean(observed - predicted)
}
\examples{
o<-1:5
p<-c(2,2,4,3,5)
me(observed=o, predicted=p)

}
